backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
cerebras.cloud.sdk==1.50.1
pymongo==4.6.0
